Move projectForm in ui-project.js to form.js
